+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v2.0.0
+## New Features
+- Add copy today stats button
+- Abstract backend electron from the application for web portability.
+- Added write mode to update deal TP. Future possibilities coming.
+- Added a Open in 3C button to deals and bots.
+- Added API error banner to display instead of the alert.
+- Added Snapshot dashboard that allows a combination of currencies, accounts, and displays statistics centered around profit.
+- Implemented the ability to show/hide specific columns.
+- Added UST
+
+## Enhancements
+- Improved the 3C API handler.
+- Reworked entire data storage logic to break profiles into set databases.
+- Improved table layout to flex, should be more stable on all screen sizes. 
+## Bug Fixes
+- Max Active Deals column was not editable on the bot planner
+- Fixed notification settings that did not persist on reload.
+- Bot API calls only pulled 100 bots at a time, improved this to pull all bots.
+- Fix tooltips that were hidden behind elements.
+- Improved reliability of the auto refresh function.
+- Multiple profiles would rewrite old deals causing a loss of data. 
+- Fixed reserved funds storing as a string instead of number.
+
 # v1.0.0
 
 ## New Features